requirements.txt
drf-spectacular>=0.28.0
drf-spectacular-sidecar>=2025.6.1

django-cors-headers>=3.15.0
